Move RetryUntilTimeout into helper/tests package
