+++ e2e/broadcast_test.go
@@ -105,7 +105,7 @@ func TestBroadcast(t *testing.T) {
 						assert.NoError(t, err)
 						assert.Equal(t, uint64(1), res.Length)
 					} else {
-						assert.ErrorIs(t, err, framework.ErrTimeout)
+						assert.ErrorIs(t, err, tests.ErrTimeout)
 					}
 				})
 			}
