+++ e2e/framework/testserver.go
@@ -21,6 +21,7 @@ import (
 	"github.com/0xPolygon/polygon-sdk/command/server"
 	"github.com/0xPolygon/polygon-sdk/consensus/ibft"
 	"github.com/0xPolygon/polygon-sdk/crypto"
+	"github.com/0xPolygon/polygon-sdk/helper/tests"
 	"github.com/0xPolygon/polygon-sdk/network"
 	"github.com/0xPolygon/polygon-sdk/server/proto"
 	txpoolProto "github.com/0xPolygon/polygon-sdk/txpool/proto"
@@ -37,10 +38,6 @@ type TestServerConfigCallback func(*TestServerConfig)
 const (
 	initialPort   = 12000
 	polygonSDKCmd = "polygon-sdk"
-)
-
-var (
-	ErrTimeout = errors.New("timeout")
 )
 
 type TestServer struct {
@@ -261,7 +258,7 @@ func (t *TestServer) Start(ctx context.Context) error {
 		return err
 	}
 
-	_, err := RetryUntilTimeout(ctx, func() (interface{}, bool) {
+	_, err := tests.RetryUntilTimeout(ctx, func() (interface{}, bool) {
 		ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
 		defer cancel()
 
@@ -362,7 +359,7 @@ func (t *TestServer) WaitForReceipt(ctx context.Context, hash web3.Hash) (*web3.
 		err     error
 	}
 
-	res, err := RetryUntilTimeout(ctx, func() (interface{}, bool) {
+	res, err := tests.RetryUntilTimeout(ctx, func() (interface{}, bool) {
 		receipt, err := client.Eth().GetTransactionReceipt(hash)
 		if err != nil && err.Error() != "not found" {
 			return result{receipt, err}, false
@@ -381,7 +378,7 @@ func (t *TestServer) WaitForReceipt(ctx context.Context, hash web3.Hash) (*web3.
 
 func (t *TestServer) WaitForReady(ctx context.Context) error {
 	client := t.JSONRPC()
-	_, err := RetryUntilTimeout(ctx, func() (interface{}, bool) {
+	_, err := tests.RetryUntilTimeout(ctx, func() (interface{}, bool) {
 		num, err := client.Eth().BlockNumber()
 		if err != nil {
 			return nil, true
